vastly improve performance
The crux of the performance issue was that I was doing a transaction
per insert. Making the bulk insert 1 transaction dramatically
improved performance. Then with some further limited benchmarking
I concluded that async insertions are unnecessary and only increased
the variance in user time of the program. This also allows me to
drop the dependency on async.